Refactor mamba wrapper
