+++ bayesflow/wrappers/__init__.py
@@ -1,1 +1,1 @@
-from .mamba import MambaSSM
+from .mamba import Mamba
