+++ bayesflow/wrappers/mamba/__init__.py
@@ -1,0 +1,1 @@
+from mamba import Mamba
